+++ src/Kalitka/Program.cs
@@ -158,6 +158,7 @@ builder.Services.AddSingleton<PolicyService>();
 builder.Services.AddSingleton<PolicyCopilot>();
 builder.Services.AddSingleton<PrincipalService>();
 builder.Services.AddSingleton<CatalogService>();
+builder.Services.AddSingleton<MyAccessService>();
 
 var app = builder.Build();
 var options = app.Services.GetRequiredService<IOptions<GateOptions>>().Value;